+++ db/migrate/20170126183101_add_hosting_request_id_to_notifications.rb
@@ -1,0 +1,6 @@
+class AddHostingRequestIdToNotifications < ActiveRecord::Migration[5.0]
+  def change
+    add_column :notifications, :hosting_request_id, :integer
+    add_index :notifications, :hosting_request_id
+  end
+end