+++ test/controllers/notifications_controller_test.rb
@@ -38,6 +38,14 @@ class NotificationsControllerTest < ActionController::TestCase
     assert_redirected_to notifications(:topic_reply_one).reply
   end
 
+  test 'should show hosting request notification' do
+    login(@regular_user)
+    get :show, params: { id: notifications(:hosting_request_one) }
+    assert_not_nil assigns(:notification)
+    assert_equal true, assigns(:notification).read
+    assert_redirected_to notifications(:hosting_request_one).hosting_request
+  end
+
   test 'should show blank notification and redirect' do
     login(@regular_user)
     get :show, params: { id: notifications(:blank) }